Added how to connect to godzilla
The tutorial is meant to be be mostly done on godzilla. With the
new restrictions in place, people will need to have set up an
ssh key pair. I've added instructions to the beginning of the
tutorial and also removed from 05 how to add a public key to
godzilla, because that does not work anymore.